Move surviveler pkg at server root
