+++ src/server/surviveler/assets_load.go
