+++ src/server/surviveler/assets_types.go
