+++ src/server/surviveler/common.go
